+++ package.json
@@ -40,7 +40,6 @@
     "@ckeditor/ckeditor5-table": "^41.3.1",
     "@ckeditor/ckeditor5-typing": "^41.3.1",
     "@ckeditor/ckeditor5-widget": "^41.3.1",
-    "@lawmatics/ckeditor5-paragraph-indent": "^1.0.4",
     "dompurify": "^3.1.3",
     "sanitize-html": "^2.13.0"
   },
